moved  controller end routine behind return of curses wrapper to ensure cleanup in any case

+++ src/controller/workerprocess/workerprocess.py
@@ -69,6 +69,8 @@ class WorkerProcess():
         return content,path.basename(filename.decode())
     
     def kill(self):
+        if self.shm == None:
+            return
         try:
             self.shm.unlink()
         except FileNotFoundError:
